-  create extractors as separate classes
